requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
redis==5.0.7
openai==1.51.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
httpx==0.27.2
